requirements.txt
openai==1.93.0
pandas==2.2.3
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.2
Pillow==10.3.0
boto3==1.40.37